arch: split dto.rs (1159 lines) into sub-modules by aggregate boundary (#294)
Split infra/dto.rs into five focused modules:
- dto/bus.rs — bus wire format, domain↔bus conversions, event serialization
- dto/task.rs — task storage DTOs
- dto/instance.rs — SM instance storage DTOs
- dto/config.rs — YAML config parsing types (agent, model, transition)
- dto/context.rs — context config and persistence types

Re-exports in dto/mod.rs preserve all existing import paths.
